+++ example/example.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import styled, { injectGlobal, keyframes } from '..'
+import styled, { createGlobalStyle, keyframes } from '..'
 
 export default () => {
-  injectGlobal`
+  const GlobalStyle = createGlobalStyle`
     body {
       font-family: sans-serif;
     }
@@ -28,6 +28,7 @@ export default () => {
     render() {
       return (
         <Wrapper>
+          <GlobalStyle />
           <Title>Hello World, this is my first styled component!</Title>
         </Wrapper>
       )