+++ src/base.js
@@ -1,16 +1,18 @@
 // @flow
 
 /* Import singletons */
+import flatten from './utils/flatten'
 import stringifyRules from './utils/stringifyRules'
 import isStyledComponent from './utils/isStyledComponent'
 import generateAlphabeticName from './utils/generateAlphabeticName'
 import css from './constructors/css'
+import _ComponentStyle from './models/ComponentStyle'
 import ServerStyleSheet from './models/ServerStyleSheet'
 import StyleSheetManager from './models/StyleSheetManager'
 
 /* Import singleton constructors */
 import _keyframes from './constructors/keyframes'
-import _injectGlobal from './constructors/injectGlobal'
+import _createGlobalStyle from './constructors/createGlobalStyle'
 
 /* Import components */
 import ThemeProvider from './models/ThemeProvider'
@@ -58,17 +60,27 @@ if (
   window['__styled-components-init__'] += 1
 }
 
+/* Instantiate internal singletons */
+const ComponentStyle = _ComponentStyle(
+  generateAlphabeticName,
+  flatten,
+  stringifyRules
+)
+
 /* Instantiate exported singletons */
 const keyframes = _keyframes(generateAlphabeticName, stringifyRules, css)
-const injectGlobal = _injectGlobal(stringifyRules, css)
-
+const createGlobalStyle = _createGlobalStyle(
+  ComponentStyle,
+  stringifyRules,
+  css
+)
 /* Export everything */
 
 export * from './secretInternals'
 export {
   css,
   keyframes,
-  injectGlobal,
+  createGlobalStyle,
   isStyledComponent,
   ThemeProvider,
   withTheme,