+++ src/models/ComponentStyle.js
@@ -1,38 +1,11 @@
 // @flow
 import hashStr from '../vendor/glamor/hash'
-
+import isStaticRules from '../utils/isStaticRules'
 import type { RuleSet, NameGenerator, Flattener, Stringifier } from '../types'
 import StyleSheet from './StyleSheet'
 import { IS_BROWSER } from '../constants'
-import isStyledComponent from '../utils/isStyledComponent'
 
 const areStylesCacheable = IS_BROWSER
-
-const isStaticRules = (rules: RuleSet, attrs?: Object): boolean => {
-  for (let i = 0, len = rules.length; i < len; i += 1) {
-    const rule = rules[i]
-
-    // recursive case
-    if (Array.isArray(rule) && !isStaticRules(rule)) {
-      return false
-    } else if (typeof rule === 'function' && !isStyledComponent(rule)) {
-      // functions are allowed to be static if they're just being
-      // used to get the classname of a nested styled component
-      return false
-    }
-  }
-
-  if (attrs !== undefined) {
-    // eslint-disable-next-line guard-for-in, no-restricted-syntax
-    for (const key in attrs) {
-      if (typeof attrs[key] === 'function') {
-        return false
-      }
-    }
-  }
-
-  return true
-}
 
 const isHMREnabled =
   process.env.NODE_ENV !== 'production' &&
@@ -88,7 +61,6 @@ export default (
 
       const flatCSS = flatten(this.rules, executionContext, styleSheet)
       const name = generateRuleHash(this.componentId + flatCSS.join(''))
-
       if (!styleSheet.hasNameForId(componentId, name)) {
         styleSheet.inject(
           this.componentId,