+++ src/test/styles.test.js
@@ -5,11 +5,9 @@ import TestRenderer from 'react-test-renderer'
 import * as nonce from '../utils/nonce'
 import { resetStyled, expectCSSMatches } from './utils'
 import StyleSheet from '../models/StyleSheet'
-import _injectGlobal from '../constructors/injectGlobal'
 import stringifyRules from '../utils/stringifyRules'
 import css from '../constructors/css'
 
-const injectGlobal = _injectGlobal(stringifyRules, css)
 
 jest.mock('../utils/nonce')
 jest.spyOn(nonce, 'default').mockImplementation(() => 'foo')