+++ src/test/utils.js
@@ -4,6 +4,7 @@
  * our public API works the way we promise/want
  */
 import _styled from '../constructors/styled'
+import _createGlobalStyle from '../constructors/createGlobalStyle'
 import css from '../constructors/css'
 import _constructWithOptions from '../constructors/constructWithOptions'
 import StyleSheet from '../models/StyleSheet'
@@ -47,13 +48,28 @@ export const resetStyled = (isServer: boolean = false) => {
   return _styled(StyledComponent, constructWithOptions)
 }
 
-const stripComments = (str: string) => str.replace(/\/\*.*?\*\/\n?/g, '')
+export const resetCreateGlobalStyle = () => {
+  const ComponentStyle = _ComponentStyle(classNames, flatten, stringifyRules)
+  return _createGlobalStyle(ComponentStyle, stringifyRules, css)
+}
+
+export const stripComments = (str: string) => str.replace(/\/\*.*?\*\/\n?/g, '')
 
 export const stripWhitespace = (str: string) =>
   str
     .trim()
     .replace(/([;\{\}])/g, '$1  ')
     .replace(/\s+/g, ' ')
+
+
+export const getCSS = (scope: Document | HTMLElement) => {
+  return Array.from(scope.querySelectorAll('style'))
+    .map(tag => tag.innerHTML)
+    .join('\n')
+    .replace(/ {/g, '{')
+    .replace(/:\s+/g, ':')
+    .replace(/:\s+;/g, ':;')
+}
 
 export const expectCSSMatches = (
   _expectation: string,
@@ -65,12 +81,7 @@ export const expectCSSMatches = (
     .replace(/:\s+/g, ':')
     .replace(/:\s+;/g, ':;')
 
-  const css = Array.from(document.querySelectorAll('style'))
-    .map(tag => tag.innerHTML)
-    .join('\n')
-    .replace(/ {/g, '{')
-    .replace(/:\s+/g, ':')
-    .replace(/:\s+;/g, ':;')
+  const css = getCSS(document)
 
   if (opts.ignoreWhitespace) {
     const stripped = stripWhitespace(stripComments(css))