+++ src/types.js
@@ -19,6 +19,10 @@ export type Target = string | ComponentType<*>
 
 export type NameGenerator = (hash: number) => string
 
+export type CSSConstructor = (
+  strings: Array<string>,
+  ...interpolations: Array<Interpolation>
+) => RuleSet
 export type StyleSheet = {
   create: Function,
 }